Added Result Variables (for every line) and Answer Variable (references the previous result) Added custom lookup Added Variable/Function sidebar Added "Save as Variable" and "Save as Function"
Entries in Variable list are clickable to insert into current formula. Variable ordering in functions is (semi-)random (/dependent on Set<String> internal order).
Terms do not recalculate when variables/functions change! This is a feature for the future, however be careful, this could result in evaluation loops (x is set to 4, x+1 results in 5, this result is saved in x, therefore x is set to 5, x+1 results in 6, ...)
Logic for overwriting functions is not yet airtight. What happens with overwritten functions in already calculated terms? Recalculate with new formula? (Attention to evaluation loops)